fix transactions to flaot

+++ src/Entity/Transaction.php
@@ -120,12 +120,12 @@ class Transaction
         return $this;
     }
 
-    public function getAmount(): int
+    public function getAmount(): float
     {
         return $this->amount;
     }
 
-    public function setAmount(int $amount): self
+    public function setAmount(float $amount): self
     {
         $this->amount = $amount;
